Preliminary Lava design working

+++ src/game.cpp
@@ -82,7 +82,7 @@ void Game::Update() {
   snake.Update();
 
   if (snake.alive && lava!=nullptr) {
-    SDL_Point head_cell{snake.head_x, snake.head_y};
+    SDL_Point head_cell{static_cast<int>(snake.head_x), static_cast<int>(snake.head_y)};
     if(lava->checkPointCollision(lava->lava_body, head_cell)) {
       snake.alive = false;
     }
